# Speed up instance variable cache misses