--- a/vm.c
+++ b/vm.c
@@ -633,6 +633,8 @@ rb_dtrace_setup(rb_execution_context_t *ec, VALUE klass, ID id,
     return FALSE;
 }
 
+extern unsigned int redblack_buffer_size;
+
 /*
  *  call-seq:
  *    RubyVM.stat -> Hash
@@ -660,6 +662,7 @@ static VALUE
 vm_stat(int argc, VALUE *argv, VALUE self)
 {
     static VALUE sym_constant_cache_invalidations, sym_constant_cache_misses, sym_global_cvar_state, sym_next_shape_id;
+    static VALUE sym_shape_cache_size;
     VALUE arg = Qnil;
     VALUE hash = Qnil, key = Qnil;
 
@@ -681,6 +684,7 @@ vm_stat(int argc, VALUE *argv, VALUE self)
     S(constant_cache_misses);
         S(global_cvar_state);
     S(next_shape_id);
+    S(shape_cache_size);
 #undef S
 
 #define SET(name, attr) \
@@ -693,6 +697,7 @@ vm_stat(int argc, VALUE *argv, VALUE self)
     SET(constant_cache_misses, ruby_vm_constant_cache_misses);
     SET(global_cvar_state, ruby_vm_global_cvar_state);
     SET(next_shape_id, (rb_serial_t)GET_SHAPE_TREE()->next_shape_id);
+    SET(shape_cache_size, (rb_serial_t)GET_SHAPE_TREE()->cache_size);
 #undef SET
 
 #if USE_DEBUG_COUNTER
@@ -3069,7 +3074,8 @@ vm_memsize(const void *ptr)
         vm_memsize_builtin_function_table(vm->builtin_function_table) +
         rb_id_table_memsize(vm->negative_cme_table) +
         rb_st_memsize(vm->overloaded_cme_table) +
-        vm_memsize_constant_cache()
+        vm_memsize_constant_cache() +
+        GET_SHAPE_TREE()->cache_size * sizeof(redblack_node_t)
     );
 
     // TODO