… test/ruby/test_shapes.rb
… variable.c