This PR speeds up instance variable cache misses by introducing a red-black tree as a shape cache.  With the red-black tree, we can easily check whether an instance variable has been set and what index the IV uses.  Before this change, in the worst case, IV lookup would be `O(n)`, but with the red-black tree, the worst case is `O(log n)` (where `n` == shape depth).

I added a benchmark to show the difference:

```
$ make benchmark ITEM=vm_ivar_ic_miss
/Users/aaron/.rubies/arm64/ruby-trunk/bin/ruby --disable=gems -rrubygems -I./benchmark/lib ./benchmark/benchmark-driver/exe/benchmark-driver \
	            --executables="compare-ruby::/Users/aaron/.rubies/arm64/ruby-trunk/bin/ruby --disable=gems -I.ext/common --disable-gem" \
	            --executables="built-ruby::./miniruby -I./lib -I. -I.ext/common  ./tool/runruby.rb --extout=.ext  -- --disable-gems --disable-gem" \
	            --output=markdown --output-compare -v $(find ./benchmark -maxdepth 1 -name 'vm_ivar_ic_miss' -o -name '*vm_ivar_ic_miss*.yml' -o -name '*vm_ivar_ic_miss*.rb' | sort) 
compare-ruby: ruby 3.3.0dev (2023-10-23T15:37:50Z master 62c674f98c) [arm64-darwin23]
built-ruby: ruby 3.3.0dev (2023-10-23T22:56:14Z rb-shape-index 4a7b24be69) [arm64-darwin23]
# Iteration per second (i/s)

|                 |compare-ruby|built-ruby|
|:----------------|-----------:|---------:|
|vm_ivar_ic_miss  |      3.465M|   23.031M|
|                 |           -|     6.65x|
```

This change has an impact on YJIT as well.  When IV sites become megamorphic, rather than exit, the JIT will generate machine code that does a "slow path" read on the IV.  Below is a benchmark to demonstrate:

```ruby
class Foo
  def initialize idx
    case idx
    when 0; then @c0 = 1
    when 1; then @c1 = 1
    when 2; then @c2 = 1
    when 3; then @c3 = 1
    when 4; then @c4 = 1
    when 5; then @c5 = 1
    when 6; then @c6 = 1
    when 7; then @c7 = 1
    when 8; then @c8 = 1
    end

    @a0 = @a1 = @a2 = @a3 = @a4 = @a5 = @a6 = @a7 = @a8 = @a9 = @a10 = @a11 = @a12 = @a13 = @a14 = @a15 = @a16 = @a17 = @a18 = @a19 = @a20 = @a21 = @a22 = @a23 = @a24 = @a25 = @a26 = @a27 = @a28 = @a29 = @a30 = @a31 = @a32 = @a33 = @a34 = @a35 = @a36 = @a37 = @a38 = @a39 = @a40 = @a41 = @a42 = @a43 = @a44 = @a45 = @a46 = @a47 = @a48 = @a49 = @a50 = @a51 = @a52 = @a53 = @a54 = @a55 = @a56 = @a57 = @a58 = @a59 = @a60 = @a61 = @a62 = @a63 = @a64 = @a65 = @a66 = @a67 = @a68 = @a69 = @a70 = @a71 = @a72 = @a73 = @a74 = @b = 1
  end

  def b; @b; end
end

# Force the `@b` site to be megamorphic
9.times { Class.new(Foo).new(_1).b }
obj = Foo.new(8)
60000000.times { obj.b }
```

With YJIT on master:

```
$ time ruby --yjit --yjit-call-threshold=2 -v test.rb
ruby 3.3.0dev (2023-10-23T15:37:50Z master 62c674f98c) +YJIT [arm64-darwin23]

________________________________________________________
Executed in   10.37 secs    fish           external
   usr time   10.28 secs   30.29 millis   10.25 secs
   sys time    0.13 secs    5.12 millis    0.12 secs
```

With YJIT on this branch:

```
$ time ./ruby --yjit --yjit-call-threshold=2 -v test.rb
ruby 3.3.0dev (2023-10-23T22:56:14Z rb-shape-index 4a7b24be69) +YJIT [arm64-darwin23]

________________________________________________________
Executed in    2.61 secs    fish           external
   usr time    2.45 secs    0.10 millis    2.45 secs
   sys time    0.04 secs    2.15 millis    0.04 secs
```