@byroot 2023-10-24
So the implementation goes over my head for the most part, but a couple things.

If cache misses have decent complexity now, should we bump or remove the `SHAPE_MAX_NUM_IVS` constant? https://github.com/ruby/ruby/blob/c44d65427e639ca45c17b1c9dece798bae59f047/shape.h#L34. It was added to avoid to factorial complexity situation when setting many ivars, should be resolved right?

This significantly speedup looking up for a shape ancestor, but do you think https://github.com/ruby/ruby/pull/8650 still make sense assuming redblack-tree is merged? I'd think it would for the supposedly common "close ancestor" case.

On another node I triggered an internal Shopify CI build to see if it spots any issue with your PR.