@tenderlove 2023-10-24
> If cache misses have decent complexity now, should we bump or remove the `SHAPE_MAX_NUM_IVS` constant?

We can just remove it.  It's not even used anymore, except in tests.  There's a bunch of other stuff I want to clean up after this lands, but I wanted to keep the diff as small as possible (I realize this is a big PR 😢 )