@jhawthorn 2023-10-24
> The RB tree seems to only pay off when we need to examine more than ~10 shapes. I think "near shape" hints would compliment this well.

I'm not sure. The bigger benefit from that post-this-change I think would be Jean's change to avoid flip-flopping the cache rather than the speed of ivar lookup.