@byroot 2023-10-24
> the speed of ivar lookup.

Well, the assumption is that in the rather common case of the "memoization" pattern, the common ancestor is very close, so in many case we might get away with just checking two or three shapes.

It's also very helpful for objects that are sometimes frozen sometimes not like `Set`. 

I'll try to find some time to benchmark this again now that the index was merged, see if I can demonstrate a substantial perf gain.