@casperisfine 2023-10-25
Alright, so something I totally missed until now is that all shapes with more than `ANCESTOR_CACHE_THRESHOLD` ivars do have an index, somehow I thought only one out of `ANCESTOR_CACHE_THRESHOLD` shapes had an index.

So it kinda changes things.

Still, I benchmarked my (unmodified) branch against newest master, and it's still quite significantly faster on the "close ancestor" case, which I suspect is common both in case of light memoization use and use case of frozen objects.

```
/usr/bin/ruby --disable=gems -rrubygems -I./benchmark/lib ./benchmark/benchmark-driver/exe/benchmark-driver \
	            --executables="compare-ruby::../miniruby-master -I.ext/common --disable-gem" \
	            --executables="built-ruby::./miniruby --disable-gem" \
	            --output=markdown --output-compare -v $(find ./benchmark -maxdepth 1 -name 'vm_ivar_memoize' -o -name '*vm_ivar_memoize*.yml' -o -name '*vm_ivar_memoize*.rb' | sort) 
compare-ruby: ruby 3.3.0dev (2023-10-25T07:50:00Z master 526292d9fe) [arm64-darwin22]
last_commit=LLDB: Use `expression` to save the result into the history [ci skip]
built-ruby: ruby 3.3.0dev (2023-10-25T08:53:12Z shapes_double_sear.. 63014a9464) [arm64-darwin22]
warming up......
# Iteration per second (i/s)

|                                     |compare-ruby|built-ruby|
|:------------------------------------|-----------:|---------:|
|vm_ivar_stable_shape                 |     11.145M|   11.714M|
|                                     |           -|     1.05x|
|vm_ivar_memoize_unstable_shape       |      7.534M|    9.818M|
|                                     |           -|     1.30x|
|vm_ivar_memoize_unstable_shape_miss  |     11.583M|   11.011M|
|                                     |       1.05x|         -|
|vm_ivar_unstable_undef               |      8.546M|    8.176M|
|                                     |       1.05x|         -|
|vm_ivar_divergent_shape              |      7.899M|  863.465k|
|                                     |       9.15x|         -|
|vm_ivar_divergent_shape_imbalanced   |     10.108M|    2.271M|
|                                     |       4.45x|         -|
```

So I think it could make sense to check a small number of ancestors before giving up and using the index.